rename 'import_kits' file
